Move lower-level interface functions to a separate object
