+++ tests/player_input.rs
@@ -13,7 +13,7 @@ fn integration_player_input() -> Result<(), Box<dyn Error>> {
     common::with_rocket_league(|| {
         let rlbot = rlbot::init()?;
 
-        rlbot.start_match(common::one_player_match())?;
+        rlbot.interface.start_match(common::one_player_match())?;
         rlbot.wait_for_match_start()?;
 
         let mut packeteer = rlbot.packeteer();
@@ -23,7 +23,7 @@ fn integration_player_input() -> Result<(), Box<dyn Error>> {
             Throttle: 1.0,
             ..Default::default()
         };
-        rlbot.update_player_input(input, 0)?;
+        rlbot.interface.update_player_input(input, 0)?;
 
         thread::sleep(Duration::from_secs(1));
         let end = packeteer.next()?;
