+++ tests/set_game_state_flatbuffer.rs
@@ -14,12 +14,14 @@ fn integration_set_game_state() -> Result<(), Box<dyn Error>> {
 
         #[allow(deprecated)]
         {
-            rlbot.start_match(common::one_player_match())?;
+            rlbot.interface.start_match(common::one_player_match())?;
         }
         rlbot.wait_for_match_start()?;
 
         let desired_state = teleport_to_sky();
-        rlbot.set_game_state(desired_state.finished_data())?;
+        rlbot
+            .interface
+            .set_game_state(desired_state.finished_data())?;
 
         // Sometimes setting the state takes a few frames, so wait a bit.
         thread::sleep(Duration::from_millis(100));
